requirements.txt
streamlit>=1.32.0
streamlit-webrtc>=0.47.0
groq>=0.9.0
edge-tts>=6.1.0
pydub>=0.25.1
av>=12.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
